fill component interface rename models folder to components
