+++ src/bc/rest/components/ControllerClassInterface.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace bc\rest\models;
+namespace bc\rest\components;
 
 
 interface ControllerClassInterface extends ComponentInterface {
