--- src/bc/rest/generator/ProcessorInterface.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace bc\rest\generator;
-
-interface ProcessorInterface {
-
-    /**
-     * Process openapi specs
-     *
-     * @param array $specs parsed openapis specs
-     */
-    public function processSpecs($specs);
-
-}
