--- src/bc/rest/models/ComponentInterface.php
@@ -1,9 +1,0 @@
-<?php
-
-
-namespace bc\rest\models;
-
-
-interface ComponentInterface {
-
-}
